Cleanup of HttpClient
There were several instances of polling/delay that have been replaced with responsible task blocking/unblocking. It's every bit as thread-unsafe as the rest of the class--there is a single instance property that holds the blocked task handle, not separate binary semaphores for each blocking operation--but hey, it's a single-threaded deal, and it will work nicely with the RTOS and other tasks that have things to do during HTTP operations.